requirements.txt
django>=4.2.16,<4.3.0
djangorestframework
pyyaml
requests
django-cors-headers